# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Bookstore API",
        "description": "這是書店的 API 文件",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/Register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "使用者註冊",
                "operationId": "ba577e61b16ce424e83adf7bf480c44c",
                "requestBody": {
                    "description": "註冊所需的使用者資料",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "account",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "Administrator"
                                    },
                                    "account": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "註冊成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "註冊成功"
                                        },
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "驗證錯誤"
                    }
                }
            }
        },
        "/api/auth/Login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "使用者登入",
                "operationId": "1d5d42c51423c44d1486c7c5cee8b4f7",
                "requestBody": {
                    "description": "登入所需的帳號與密碼",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "account",
                                    "password"
                                ],
                                "properties": {
                                    "account": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "登入成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "登入成功"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "帳號或密碼錯誤"
                    }
                }
            }
        },
        "/api/auth/Logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "使用者登出",
                "operationId": "5bbc0da1ea2cecb59fbb05f11195108d",
                "responses": {
                    "200": {
                        "description": "已登出",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "已登出"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "登出失敗"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}